refactor: streamline registry manager exports
