+++ src/index.ts
@@ -1,30 +1,3 @@
 export { cloneNavigationState } from './features/navigation/cloneNavigationState';
 export { createNavigation, initialState } from './navigationFactory';
-export { createRegistry } from './registry';
-export {
-  createRegistryManager,
-  setRegistry,
-  initRegistry,
-  getRegistry,
-  resetRegistry,
-  getCommands,
-  getProcessors,
-  getSources,
-  getStores,
-} from './registryManager';
-
-export * from './commands';
-export * from './processors';
-export * from './sources';
-export * from './stores';
-export type {
-  Command,
-  CliCommand,
-  VoiceCommand,
-  Processor,
-  GroupedProcessor,
-  Source,
-  Store,
-  SupabaseConfig,
-  AnalyticsConfig,
-} from './interfaces';
+export * from './registryManager';
